+++ internal/app/app.go
@@ -57,6 +57,8 @@ type WAClient interface {
 	RequestHistorySyncOnDemand(ctx context.Context, lastKnown types.MessageInfo, count int) (types.MessageID, error)
 	Logout(ctx context.Context) error
 	LinkedJID() string
+
+	SetProfilePicture(ctx context.Context, avatar []byte) (string, error)
 }
 
 type Options struct {